Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Ludorif",email = "ludorif@test.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "requests (>=2.33.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "google-genai (>=1.69.0,<2.0.0)",
    "pexels-api (>=1.0.1,<2.0.0)",
    "google-api-python-client (>=2.193.0,<3.0.0)",
    "magic-hour (>=0.58.0,<0.59.0)",
    "boto3 (>=1.42.78,<2.0.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "mypy (>=1.20.0,<2.0.0)",
    "PyJWT (>=2.10.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
